requirements.txt
alabaster==0.7.10
appdirs==1.4.3
astroid==1.6.5
autoflake==0.6.6
autopep8==1.3.5
Babel==2.6.0
bandit==1.4.0
certifi==2018.4.16
chardet==3.0.4
click==6.6
cmakelint==1.3.4.1
coala==0.11.0
coala-bears==0.11.1
coala-utils==0.6.7
colorama==0.3.9
colorlog==2.10.0
cppclean==0.12
cpplint==1.3.0
decorator==4.3.0
dennis==0.9
dependency-management==0.4.0
docopt==0.6.2
docutils==0.14
eradicate==0.1.6
Flask==1.0.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
future==0.16.0
gitdb2==2.0.3
GitPython==2.1.10
guess-language-spirit==0.5.3
html-linter==0.3.0
idna==2.7
imagesize==1.0.0
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
lazy-object-proxy==1.3.1
libclang-py3==3.4.0
mando==0.3.3
MarkupSafe==1.0
mccabe==0.6.1
munkres3==1.0.5.5
mypy-lang==0.4.6
nbformat==4.4.0
nltk==3.3
numpy==1.14.3
packaging==17.1
pbr==4.0.4
Pillow==5.0.0
polib==1.1.0
proselint==0.7.0
pycodestyle==2.4.0
pydocstyle==1.1.1
pyflakes==1.4.0
Pygments==2.2.0
pylint==1.9.2
pyparsing==2.2.0
PyPrint==0.2.6
pyroma==2.2
pytz==2018.4
PyYAML==3.12
radon==1.4.0
requests==2.19.0
restructuredtext-lint==1.0.1
rstcheck==2.2
safety==0.5.1
sarge==0.1.4
scspell3k==2.1
six==1.11.0
smmap2==2.0.3
snowballstemmer==1.2.1
Sphinx==1.4.9
stevedore==1.28.0
template-remover==0.1.9
termcolor==1.1.0
traitlets==4.3.2
unidiff==0.5.5
urllib3==1.23
vulture==0.10
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
yamllint==1.6.1
yapf==0.14.0
